services/ingest_api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
google-cloud-storage==2.18.2
google-cloud-firestore==2.16.0
google-cloud-pubsub==2.21.5
google-cloud-documentai==3.6.0
cloudevents==1.10.1
